@@ -11,18 +11,17 @@
 	"author": "",
 	"license": "PRIVATE",
 	"dependencies": {
-		"bcrypt": "^3.0.5",
-		"body-parser": "^1.18.3",
+		"bcrypt": "^3.0.6",
 		"express": "^4.16.4",
-		"express-session": "^1.15.6",
-		"nodemailer": "^6.1.0",
-		"pg": "^7.9.0",
+		"express-session": "^1.16.1",
+		"nodemailer": "^6.1.1",
+		"pg": "^7.10.0",
 		"random": "^2.1.1"
 	},
 	"devDependencies": {
 		"@types/express": "^4.16.1",
-		"@types/node": "^11.12.0",
-		"typescript": "^3.3.4000"
+		"@types/node": "^11.13.8",
+		"typescript": "^3.4.5"
 	},
 	"engines": {
 		"node": "11.x"
