@@ -9,7 +9,8 @@ import { Observable } from 'rxjs';
   styleUrls: ['./contact.component.css']
 })
 export class ContactComponent implements OnInit {
-  
+
+    sending : boolean = false;
     form : EmailForm;
     error : string = "";
 
@@ -31,7 +32,7 @@ export class ContactComponent implements OnInit {
 
         if (this.form.address == "") {
           this.error = "Introduzca una dirección de correo electrónico, por favor.";
-          return; 
+          return;  
         }
         if (this.form.subject == "") {
           this.error = "Introduzca el asunto de su mensaje, por favor.";
@@ -42,15 +43,20 @@ export class ContactComponent implements OnInit {
           console.log(this.form.body);
           return;
         } 
-
-        this.dataService.sendContactEmail(this.form).subscribe((data) => {
-          if (data.ok) {
-            alert("Su mensaje ha sido enviado con éxito. El administrador se pondrá en contacto usted lo antes posible.");
-            window.location.href = "/";
-          } else {
-            alert("Su mensaje no ha podido ser enviado. Por favor, inténtelo de nuevo más tarde.")
-          }
-        });
+        
+        if (!this.sending) {
+          this.sending = true;
+          this.dataService.sendContactEmail(this.form).subscribe((data) => {
+            if (data.ok) {
+              alert("Su mensaje ha sido enviado con éxito. El administrador se pondrá en contacto usted lo antes posible.");
+              this.sending = false;
+            } else {
+              alert("Su mensaje no ha podido ser enviado. Por favor, inténtelo de nuevo más tarde.")
+              this.sending = false;
+            }
+          });
+          
+        }
     }
 
 }
